+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.5.1-SNAPSHOT
 
+* [CHANGE] Target view shows also targets with no active alerts (targets with at least one record in journal).
 * [ENHANCEMENT] Added `clearTimestamp` and `clearUid` to alerts; timestamp when alert was cleared
 * [ENHANCEMENT] Added some javadoc